rename utils.proto
