+++ grpc/src/main/proto/choam.proto
@@ -4,7 +4,7 @@ option java_multiple_files = true;
 option java_package = "com.salesfoce.apollo.choam.proto";
 option java_outer_classname = "ChoamProto";
 option objc_class_prefix = "Chp";
-import "util.proto";
+import "crypto.proto";
 import "stereotomy.proto";
 
 package apollo.choam;
@@ -15,7 +15,7 @@ service TransactionSubmission {
 
 service Terminal {
     /* reconfiguration */
-    rpc join (utils.Digeste) returns (ViewMember) {}
+    rpc join (crypto.Digeste) returns (ViewMember) {}
 
     /* bootstrapping */
     rpc sync(Synchronize) returns (Initial) {}
@@ -55,10 +55,10 @@ message Header {
     uint64 height = 1; 
     uint64 lastCheckpoint = 2;
     uint64 lastReconfig = 3;
-    utils.Digeste lastCheckpointHash = 4;
-    utils.Digeste lastReconfigHash = 5;
-    utils.Digeste previous = 6;
-    utils.Digeste bodyHash = 7;
+    crypto.Digeste lastCheckpointHash = 4;
+    crypto.Digeste lastReconfigHash = 5;
+    crypto.Digeste previous = 6;
+    crypto.Digeste bodyHash = 7;
 }
 
 message Genesis {
@@ -68,7 +68,7 @@ message Genesis {
 }
 
 message Reconfigure {
-    utils.Digeste id = 1;
+    crypto.Digeste id = 1;
     int32 checkpointTarget = 2; 
     repeated Join joins = 3;
 }
@@ -76,8 +76,8 @@ message Reconfigure {
 message Checkpoint { 
     int64 byteSize = 1;
     int32 segmentSize = 2;
-    utils.Digeste stateHash = 3;
-    repeated utils.Digeste segments = 4;
+    crypto.Digeste stateHash = 3;
+    repeated crypto.Digeste segments = 4;
 }
 
 message Executions {
@@ -85,24 +85,24 @@ message Executions {
 }
 
 message Assemble {
-    utils.Digeste nextView = 1;
+    crypto.Digeste nextView = 1;
 }
 
 message Foundation {
-    repeated utils.Digeste membership = 1;
+    repeated crypto.Digeste membership = 1;
     stereotomy.KeyEvent_ authority = 2;
 }
 
 message FoundationSeal {
     Foundation foundation = 1;
-    utils.Sig signature = 2;
+    crypto.Sig signature = 2;
 }
 
 message Transaction { 
-    utils.Digeste source = 1;
+    crypto.Digeste source = 1;
     int32 nonce = 2;
     bytes content = 3;
-    utils.Sig signature = 4; 
+    crypto.Sig signature = 4;
 }
 
 message UnitData {
@@ -112,21 +112,21 @@ message UnitData {
 }
 
 message Join {
-    utils.Digeste view = 1;
+    crypto.Digeste view = 1;
     ViewMember member = 2;
     repeated Certification endorsements = 3;
     stereotomy.KERL_ kerl = 4;
 }
 
 message ViewMember {
-    utils.Digeste id = 1;
-    utils.PubKey consensusKey = 2;
-    utils.Sig signature = 3;
+    crypto.Digeste id = 1;
+    crypto.PubKey consensusKey = 2;
+    crypto.Sig signature = 3;
 }
 
 message Certification {
-    utils.Digeste id = 1;
-    utils.Sig signature = 2;
+    crypto.Digeste id = 1;
+    crypto.Sig signature = 2;
 }
 
 message Certifications {
@@ -147,7 +147,7 @@ message CheckpointProcessing {
 }
 
 message Validate {  
-    utils.Digeste hash = 1;
+    crypto.Digeste hash = 1;
     Certification witness = 2;
 }
 
@@ -165,7 +165,7 @@ message Validations {
 }
 
 message Synchronize {
-    utils.Digeste from = 1;
+    crypto.Digeste from = 1;
     uint64 height = 2;
 }
 
@@ -179,7 +179,7 @@ message Initial {
 message BlockReplication {
     uint64 from = 1;
     uint64 to = 2;
-    utils.Biff blocksBff = 3;
+    crypto.Biff blocksBff = 3;
 }
 
 message Blocks {
@@ -188,12 +188,12 @@ message Blocks {
 
 message CheckpointReplication {
     uint64 checkpoint = 1;
-    utils.Biff checkpointSegments = 2;
+    crypto.Biff checkpointSegments = 2;
 }
 
 message CheckpointSegments {
     repeated Slice segments = 1;
-    repeated utils.Digeste blocks = 2;
+    repeated crypto.Digeste blocks = 2;
 }
 
 message Slice {
