+++ grpc/src/main/proto/crypto.proto
@@ -1,13 +1,13 @@
 syntax = "proto3";
 
 option java_multiple_files = true;
-option java_package = "com.salesfoce.apollo.utils.proto";
+option java_package = "com.salesfoce.apollo.cryptography.proto";
 option java_outer_classname = "ApolloUtilsProto";
 option objc_class_prefix = "Au";
 import "google/protobuf/any.proto";
 import "google/protobuf/timestamp.proto";
 
-package utils;
+package crypto;
  
 message Biff {
         int32 m = 1;
