+++ grpc/src/main/proto/messaging.proto
@@ -6,13 +6,13 @@ option java_outer_classname = "MessagingProto";
 option objc_class_prefix = "Mp";
 import "google/protobuf/empty.proto";
 import "google/protobuf/any.proto";
-import "util.proto";
+import "crypto.proto";
 
 package messaging;
 
 message MessageBff {
     int32 ring = 1;
-    utils.Biff digests = 2;
+    crypto.Biff digests = 2;
 }
 
 message ByteMessage {
@@ -26,7 +26,7 @@ service RBC {
 
 message Reconcile {
     repeated  AgedMessage updates = 1;
-    utils.Biff digests = 2;
+    crypto.Biff digests = 2;
 }
 
 message ReconcileContext {
@@ -40,12 +40,12 @@ message AgedMessage {
 }
 
 message DefaultMessage {
-    utils.Digeste source = 1;
+    crypto.Digeste source = 1;
     int32 nonce = 2;
     bytes content = 3;
 }
 
 message SignedDefaultMessage {
     DefaultMessage content = 1;
-    utils.Sig signature = 2;
+    crypto.Sig signature = 2;
 }
