+++ model/src/main/java/com/salesforce/apollo/model/SubDomain.java
@@ -9,8 +9,8 @@ package com.salesforce.apollo.model;
 import com.codahale.metrics.Timer;
 import com.salesfoce.apollo.demesne.proto.DelegationUpdate;
 import com.salesfoce.apollo.demesne.proto.SignedDelegate;
-import com.salesfoce.apollo.utils.proto.Biff;
-import com.salesfoce.apollo.utils.proto.Digeste;
+import com.salesfoce.apollo.cryptography.proto.Biff;
+import com.salesfoce.apollo.cryptography.proto.Digeste;
 import com.salesforce.apollo.archipelago.Enclave.RoutingClientIdentity;
 import com.salesforce.apollo.archipelago.RouterImpl.CommonCommunications;
 import com.salesforce.apollo.choam.Parameters.Builder;
