+++ model/src/main/java/com/salesforce/apollo/model/comms/Delegation.java
@@ -8,7 +8,7 @@ package com.salesforce.apollo.model.comms;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import com.salesfoce.apollo.demesne.proto.DelegationUpdate;
-import com.salesfoce.apollo.utils.proto.Biff;
+import com.salesfoce.apollo.cryptography.proto.Biff;
 import com.salesforce.apollo.archipelago.Link;
 
 /**
