+++ model/src/main/java/com/salesforce/apollo/model/comms/DelegationService.java
@@ -7,7 +7,7 @@
 package com.salesforce.apollo.model.comms;
 
 import com.salesfoce.apollo.demesne.proto.DelegationUpdate;
-import com.salesfoce.apollo.utils.proto.Biff;
+import com.salesfoce.apollo.cryptography.proto.Biff;
 import com.salesforce.apollo.crypto.Digest;
 
 /**
