+++ model/src/main/java/com/salesforce/apollo/model/demesnes/comm/OuterContextServer.java
@@ -10,7 +10,7 @@ import com.codahale.metrics.Timer.Context;
 import com.google.protobuf.Empty;
 import com.salesfoce.apollo.demesne.proto.OuterContextGrpc.OuterContextImplBase;
 import com.salesfoce.apollo.demesne.proto.SubContext;
-import com.salesfoce.apollo.utils.proto.Digeste;
+import com.salesfoce.apollo.cryptography.proto.Digeste;
 import com.salesforce.apollo.model.comms.OuterServerMetrics;
 
 import io.grpc.stub.StreamObserver;
